Populate PackageInfo model with info from files.json

+++ conda/gateways/disk/read.py
@@ -10,6 +10,7 @@ from itertools import chain
 from logging import getLogger
 from os.path import isfile, islink, join
 
+from ..._vendor.toolz.itertoolz import pluck
 from ...base.constants import FileMode, PREFIX_PLACEHOLDER, UTF8
 from ...models.package_info import PackageInfoContents
 from ...models.record import Record
@@ -45,17 +46,36 @@ def collect_all_info_for_package(extracted_package_directory):
     info_dir = join(extracted_package_directory, 'info')
 
     files_path = join(extracted_package_directory, 'info', 'files')
-    files = tuple(ln for ln in (line.strip() for line in yield_lines(files_path)) if ln)
+    file_json_path = join(extracted_package_directory, 'info', 'files.json')
 
-    has_prefix_files = read_has_prefix(join(info_dir, 'has_prefix'))
-    no_link = read_no_link(info_dir)
-    soft_links = read_soft_links(extracted_package_directory, files)
-    index_json_record = read_index_json(extracted_package_directory)
-    icondata = read_icondata(extracted_package_directory)
-    noarch = read_noarch(extracted_package_directory)
+    if isfile(file_json_path):
+        with open(file_json_path) as file_json:
+            data = json.load(file_json)
+        data = data.get("files")
 
-    return PackageInfoContents(files, has_prefix_files, no_link, soft_links,
-                               index_json_record, icondata, noarch)
+        files = tuple(pluck("short_path", data))
+        has_prefix_files = {
+            f.get("short_path"): tuple([f.get("prefix_placeholder"), f.get("file_mode")]) for f in
+            data if f.get("prefix_placeholder") is not None}
+        no_link = set([f.get("no_link") for f in data if f.get("no_link") is not None])
+        soft_links = tuple([f for f in list(pluck("file_type", data)) if f == "softlink"])
+        index_json_record = read_index_json(extracted_package_directory)
+        icondata = read_icondata(extracted_package_directory)
+        noarch = read_noarch(extracted_package_directory)
+        return PackageInfoContents(files=files, has_prefix_files=has_prefix_files, no_link=no_link,
+                                   soft_links=soft_links, index_json_record=index_json_record,
+                                   icondata=icondata, noarch=noarch)
+    else:
+        files = tuple(ln for ln in (line.strip() for line in yield_lines(files_path)) if ln)
+
+        has_prefix_files = read_has_prefix(join(info_dir, 'has_prefix'))
+        no_link = read_no_link(info_dir)
+        soft_links = read_soft_links(extracted_package_directory, files)
+        index_json_record = read_index_json(extracted_package_directory)
+        icondata = read_icondata(extracted_package_directory)
+        noarch = read_noarch(extracted_package_directory)
+        return PackageInfoContents(files, has_prefix_files, no_link, soft_links,
+                                   index_json_record, icondata, noarch)
 
 
 def read_noarch(extracted_package_directory):
